Handle Living Graveyard sustained movement
